Moves two test helper functions into utils
