+++ tests/utils.py
@@ -89,3 +89,16 @@ def dastollervey_datasaver(func):
         return result
 
     return wrapper
+
+
+def load_auth():
+    "Helper function to load token"
+    if os.path.exists("testtoken.json"):
+        with open("testtoken.json") as fobj:
+            return json.load(fobj)
+    return None
+
+
+def save_auth(token):
+    with open("testtoken.json", "w") as fobj:
+        json.dump(token, fobj)
